Bug fix 130311: Fix the probability when moving branch.
There is used to be an array in the parent that stores the _depth of its children. When a children is removed (deleted) by removeBranch(), the array is not modified accordingly. This makes the probability assigned incorrectly to each child. This array is now removed, and its functionality is maintained since each child knows its depth already.