Move the check for error state before findSucc(). Allow Checker to function correctly.
The Checker updates the CheckerState when Checker::check() is invoked. If Checker::check() is invoked after the creation of children, the children will get a default copy of CheckerState. Moving Checker::check() before findSucc() will update the CheckerState at parent node, and this copy will be passed on the children when children are begin created in findSucc()